Refactored CustomerController
